install sidekiq

+++ start.sh
@@ -1,5 +1,8 @@
 #!/bin/bash
 set -e
-
-rm -f tmp/pids/server.pid
-rails server -b 0.0.0.0
+if [ "$WORKER" == true ]; then
+  RAILS_ENV=$RAILS_ENV bundle exec sidekiq -C config/sidekiq.yml
+else
+  rm -f tmp/pids/server.pid
+  rails server -b 0.0.0.0
+fi
